refactor file name and migrations
