+++ apps/users/models/__init__.py
@@ -1,8 +1,8 @@
 from .user import User
 from .password import SaltedPasswordModel
 
-from .nationaladmin import NationalAdmin
-from .nationalauthority import NationalAuthority
+from .national_admin import NationalAdmin
+from .national_authority import NationalAuthority
 
-from .divisionadmin import DivisionalAdmin
-from .divisionauthority import DivisionalAuthority
+from .division_admin import DivisionalAdmin
+from .division_authority import DivisionalAuthority
