+++ apps/users/views.py
@@ -1,21 +1,15 @@
-from django.shortcuts import render
 from rest_framework.views import APIView
 from rest_framework.response import Response
-from rest_framework_simplejwt.authentication import JWTAuthentication
-from apps.users.models.divisionauthority import DivisionalAuthority
-from apps.users.models.divisionadmin import DivisionalAdmin
-from apps.users.models.admin import Admin
-from apps.users.models.nationaladmin import NationalAdmin
-from apps.users.models.nationalauthority import NationalAuthority
+from apps.users.models.division_authority import DivisionalAuthority
+from apps.users.models.division_admin import DivisionalAdmin
+from apps.users.models.national_admin import NationalAdmin
+from apps.users.models.national_authority import NationalAuthority
 
-from apps.users.models.user import User
 from apps.users.serializers.user import UserSerializer
 from .serializers.admin import RegisterAdminSerializer
 from utils.email import ActivationEmail
-from django.shortcuts import render
 from rest_framework.viewsets import ModelViewSet
-from rest_framework.permissions import DjangoModelPermissions, DjangoObjectPermissions
-from django.contrib.auth.backends import ModelBackend
+from rest_framework.permissions import DjangoModelPermissions
 
 
 class CustomDjangoModelPermissions(DjangoModelPermissions):
